DCKUBE-410: Split container tests out into their own class.
